debug.c: sketch experiment with early debugging of inferior
The idea here is to install fake r_debug and link_map structures in the
created DT_DEBUG entry in the chain loader, so that debugging can work
during the early execution of the inferior ld.so. It will eventually
clobber the DT_DEBUG with its own initializers. The main piece of code
here is #ifdef'd out because it causes glibc's ld.so to crash... need to
understand why.

Conflicts:
	src/debug.c